+++ src/main/java/mops/Rheinjug2.java
@@ -8,7 +8,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 public class Rheinjug2 {
 
   public static void main(String[] args) {
-
     SpringApplication.run(Rheinjug2.class, args);
   }
