+++ src/main/resources/application.properties
@@ -15,7 +15,7 @@ keycloak.resource=demo
 keycloak.public-client=true
 spring.servlet.multipart.enabled=true
 spring.servlet.multipart.max-file-size=15MB
-spring.servlet.multipart.max-request-size=1
+spring.servlet.multipart.max-request-size=50MB
 minio.buckek.name=grossbrandtmeister
 minio.default.folder=/
 minio.access.name=Q3AM3UQ867SPQQA43P2F
